android-plugins: convert to full plugins
