--- build-logic/android-plugins/src/main/kotlin/com.github.android-password-store.android-library.gradle.kts
@@ -1,9 +1,0 @@
-/*
- * Copyright © 2014-2021 The Android Password Store Authors. All Rights Reserved.
- * SPDX-License-Identifier: GPL-3.0-only
- */
-
-plugins {
-  id("com.android.library")
-  id("com.github.android-password-store.android-common")
-}
