+++ build-logic/android-plugins/src/main/kotlin/dev/msfjarvis/aps/gradle/flavors/SlimTests.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: GPL-3.0-only
  */
 
-package flavors
+package dev.msfjarvis.aps.gradle.flavors
 
 import com.android.build.api.variant.ApplicationAndroidComponentsExtension
 import com.android.build.api.variant.LibraryAndroidComponentsExtension
